rtx.conf for gmod.exe
Garrysmod recently had an update that nuked hl2.exe. Hl2.exe was tied to a config file within Remix that enabled rtx.useObsoleteHashOnTextureUpload = True - because of this the hashes being used are now different. Going forward,  we set this to rtx.useObsoleteHashOnTextureUpload = False revealing the real hashes that belong to gmod.exe - mods will need to be updated to reflect these new hashes.